Removed all checks for SDK version, assumes 1.9.22 or later.
The SDK version 1.9.49 has a release version of 0.0.0, which makes
all version comparisons pointless. There were very few places we
were actually switching behaviour because of old versions, and any
SDK since 1.9.22 will work fine.

Because of the 1.9.49 version things, Djangae's runserver command
was raising a TypeError. This fixes it.