Renaming some Python imports to use the module.
There's a ton of things we need to import from the blobstore module,
and it is going to get longer. Instead of having very ugly / verbose
import lines, lets use the module's namespace.

This required a rename for variables in the tests, else we would
get a NameError in a lot of cases.