Added djangae.storage.parse_gs_object_name(fileobj).
This is a helper for extracting the X-AppEngine-Cloud-Storage-Object
header value from the contents of a file submitted via App Engine's
`blobstore.create_upload_url()` thingy.

You use it with the InMemoryUploadedFile instance that would be
created by the callback POST (where the user-subitted file is
replaced with a stub file which contains text about the newly-
created object in Google Cloud Storage).